requirements.txt
GitDB # TO ADD
alabaster==0.7.12
astroid==2.2.5
awscli==1.16.269
Babel==2.7.0
botocore==1.13.5
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
coverage==4.5.4
datashape==0.5.2
decorator==4.4.1
docutils==0.15.2
entrypoints==0.3
flake8==3.7.9
graphviz
GitPython==3.0.8
idna==2.8
imagesize==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
m2r==0.2.1
matplotlib==2.2.4
MarkupSafe==1.1.1
mccabe==0.6.1
multipledispatch==0.6.0
networkx==2.4
numpy==1.17.3
odo==0.5.0
packaging==19.2
pandas==0.25.2
psycopg2-binary==2.8.4
pyasn1==0.4.7
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.3
pymc3==3.7
pyparsing==2.4.2
pytest>=4.3.1
pytest-cov>=2.6.1
python-box==3.4.5
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
schema==0.7.1
six==1.12.0
scikit-learn==0.21.2
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.10
toolz==0.10.0
urllib3==1.25.6
xgboost==0.90
